pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pslipstream"
version = "1.0.0"
description = "The most informative Home-media backup solution."
license = "GPLv3"
authors = ["rlaphoenix <rlaphoenix@pm.me>"]
readme = "README.md"
repository = "https://github.com/rlaphoenix/slipstream"
keywords = ["python", "dvd", "backup"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Microsoft :: Windows :: Windows 7",
    "Operating System :: Microsoft :: Windows :: Windows 8",
    "Operating System :: Microsoft :: Windows :: Windows 8.1",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Operating System :: Microsoft :: Windows :: Windows Server 2008",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Video :: Conversion",
    "Topic :: Security :: Cryptography",
]
include = [
    { path = "CHANGELOG.md", format = "sdist" },
    { path = "README.md", format = "sdist" },
    { path = "LICENSE", format = "sdist" },
    "static/*",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/rlaphoenix/slipstream/issues"
"Forums" = "https://github.com/rlaphoenix/slipstream/discussions"
"Changelog" = "https://github.com/rlaphoenix/slipstream/blob/master/CHANGELOG.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pycdlib = "^1.14.0"
pydvdcss = "^1.4.0"
pydvdid-m = "^1.1.1"
appdirs = "^1.4.4"
tqdm = "^4.66.1"
PySide6-Essentials = "^6.5.3"
click = "^8.1.7"
coloredlogs = "^15.0.1"
jsonpickle = "^3.0.2"
pywin32 = {version = "306", platform = "win32"}
WMI = {version = "^1.5.1", platform = "win32"}
pyinstaller = {version = "^6.0.0", optional = true}

[tool.poetry.dev-dependencies]
ruff = "^0.0.292"
isort = "^5.12.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
types-python-dateutil = "^2.8.19.14"

[tool.poetry.extras]
pyinstaller = ["pyinstaller"]

[tool.poetry.scripts]
slipstream = 'pslipstream.main:main'

[tool.ruff]
exclude = [
    ".venv",
    "build",
    "dist",
    "*_pb2.py",
    "*.pyi",
]
ignore = []
line-length = 120
select = ["E", "F", "W"]

[tool.isort]
line_length = 120

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
follow_imports = 'silent'
ignore_missing_imports = true
no_implicit_optional = true
